HV: guest: cleanup of remaining integral type violations
Clean up most reported integral-type-related violations still existing under
arch/x86/guest/. The remaining reports that are not trivial to suppress will be
explained in separate documents.

Also move acpi_info outside acrn_common.h as the structure is no longer shared
with DM.

v1 -> v2:

    * Move struct acpi_info to bsp_extern.h

Acked-by: Eddie Dong <eddie.dong@intel.com>

+++ hypervisor/arch/x86/guest/instr_emul.c
@@ -1435,7 +1435,7 @@ emulate_bittest(struct vcpu *vcpu, uint64_t gpa, struct vie *vie,
 {
 	uint64_t val, rflags, bitmask;
 	int error;
-	uint32_t  bitoff;
+	uint64_t bitoff;
 	uint8_t size;
 
 	/*
@@ -1602,7 +1602,7 @@ vie_calculate_gla(enum vm_cpu_mode cpu_mode, enum cpu_reg_name seg,
 	    "%s: invalid segment %d", __func__, seg);
 	ASSERT(length == 1U || length == 2U || length == 4U || length == 8U,
 	    "%s: invalid operand size %hhu", __func__, length);
-	ASSERT((prot & ~(PROT_READ | PROT_WRITE)) == 0,
+	ASSERT((prot & ~(PROT_READ | PROT_WRITE)) == 0U,
 	    "%s: invalid prot %#x", __func__, prot);
 
 	firstoff = offset;
